New release of pipeline (#152)
* New version of pipeline invocation

* Make use of new chart version

* Creating cluster fails when using version 1.12.6

* Give error if cluster do not exist

* Make sure creating cluster works stable

* Now it appears to predictably go to success

* Make a check for the operator to be present
in cluster

* Should be marked as internal

+++ scripts/teardown_cluster.sh
@@ -102,6 +102,15 @@ fi
 ### Delete cluster
 ###
 
+# Exit if cluster does not exist
+echo ""
+echo "Connecting kubectl..."
+if [[ ""$(az aks get-credentials --overwrite-existing --admin --resource-group "$RESOURCE_GROUP"  --name "$CLUSTER_NAME" 2>&1)"" == *"ERROR"* ]]; then    
+    # Send message to stderr
+    echo -e "Error: Cluster \"$CLUSTER_NAME\" not found." >&2
+    exit 0        
+fi
+
 echo ""
 echo "Deleting cluster..."
 az aks delete --resource-group "$RESOURCE_GROUP" --name "$CLUSTER_NAME" --yes
